+++ package.json
@@ -41,6 +41,7 @@
     "webpack-dev-server": "^4.8.1"
   },
   "dependencies": {
+    "@reduxjs/toolkit": "^1.9.5",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-redux": "^7.2.4",
