+++ src/components/RegionsContainer.jsx
@@ -1,6 +1,6 @@
 import { useDispatch, useSelector } from 'react-redux';
 
-import { selectRegion, loadRestaurants } from '../redux/actions';
+import { selectRegion, loadRestaurants } from '../redux/slice';
 
 import { get } from '../utils/utils';
